refine defination of one data type in psb_drm.h
BZ:220667

some structures are shared between 64-bit and 32-bit
system, so size of member in these data structure must be
clear.

Reviewed-on: https://android.intel.com:443/238322

+++ interface/psb_drm.h
@@ -862,7 +862,7 @@ struct drm_psb_register_rw_arg {
 		uint32_t b_wms;
 		uint32_t buffer_handle;
 		uint32_t backbuf_index;
-		unsigned long backbuf_addr;
+		uint32_t backbuf_addr;
 	} overlay;
 
 	uint32_t vsync_operation_mask;
